Consolidate a bit
This is recommended practice for `apt-get install` commands
in the Docker user manual:
https://docs.docker.com/engine/userguide/eng-image/dockerfile_best-practices/#/run

But `apt-get update` is failing right now, with

  W: Failed to fetch http://deb.debian.org/debian/dists/stable/main/binary-amd64/Packages  Hash Sum mismatch

  E: Some index files failed to download. They have been ignored, or old ones used instead.

That worked the first time I ran a version of this Dockerfile less
than an hour ago, also straight out of the Docker jessie base image,
but a later `apt-get install` failed with a very similar error on a
particular package.  Transient inconsistency in the archive?  Will
have to try again later.